Safari-611.1.10/LayoutTests/fast/viewport/viewport-32-expected.txt

ALERT: viewport size 980x200 scale 1.76 with limits [1.76, 5] and userScalable true
